+++ src/components/screen/composeGame.js
@@ -1,0 +1,36 @@
+import pubSub from '../../containers/pubSub';
+
+export default (state) => ({
+  // init() {
+  //   pubSub.publish('notify', 'Place ships');
+  //   this.start = this.start.bind(this);
+  //   this.renderShip = this.renderShip.bind(this);
+  // },
+  renderShip(element) {
+    // This will append to the content div
+    console.log(element);
+  },
+  reset(e) {
+    // Clears board
+  },
+  start(e) {
+    // Set this.gameReady to true
+    // Publish something...?
+    // Reveal player two's board
+    this.gameReady = true;
+    this.render();
+    this.renderWait();
+  },
+  boardHandler(e) {
+    const btn = e.target;
+    const x = parseInt(e.target.dataset.x);
+    const y = parseInt(e.target.dataset.y);
+    if (!isNaN(x) || !isNaN(y)) {
+      // Place ship
+      console.log(`Placing a ship starting at [${x}, ${y}]`);
+      this.game.playerOneBoard.placeShip([2, 2], 3, false);
+      this.game.playerTwoBoard.placeShip([6, 2], 3, false);
+      this.renderShip(btn);
+    }
+  },
+});