+++ src/containers/isHuman.js
@@ -1,5 +1,6 @@
 export default (player) => ({
   attack: ([x, y]) => {
     player.opponentBoard.receiveAttack([x, y]);
+    return [x, y];
   },
 });